docker/consumer/requirements.txt
kafka-python
PyHDFS